Started work on setupscript
Setupscript is a Python script to set up Ruby on Rails.

Why Python? Because a capable enough version of Python is available on
all versions of OS X, Linux etc. Not so useful for Windows, but
whatever. If you are using Windows, you'll need your hand holding
anyway. This just automates things for people using Linux and OS X.

setupscript uses a bunch of stuff from buildup, a script I use privately
for doing system setup.