added assign_slots and description to timeslots
Added assign_slots boolean property to Timeslots model: this is to mark
timeslots where slots should not be generated. It defaults to true:
change to false to not assign slots for that timeslot.

Timeslots can now contain a description which is to provide
supplementary information for a timeslot that has been set to not have
assignable slots. Basically, the point of this is so that when it is
lunchtime or teatime or the Mandatory Half Hour Period Where We Must Do
Silly Walks as mandated by King George or whatever, we can provide
explanatory information in here. For instance, at BarCamp London 8, we
had two minutes of silence for Remembrance Day - we could use a
non-assignable slot for that, and put the description in as a string in
that period.